payload message dto : models id from type long to integer

+++ src/main/java/com/example/immo/dto/payloads/PayloadMessageDto.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PayloadMessageDto {
-    private Long rental_id;
-    private Long user_id;
+    private Integer rental_id;
+    private Integer user_id;
     private String message;
 }
